Removed Erlang packages as runtime dependency from Raspbian build
Since there currently is no official Erlang 18.2 package for Raspbian
that could be installed from the Debian/Raspbian repos using apt-get
having this dependency in the RVI package causes issues with dpkg on
the target.

Since we manually install the Erlang packages on the RPi using dpkg
directly the package dependency on erlang cannot be fulfilled as this
package is a meta package that only lists the other Erlang packages
as dependencies. However, meta packages only work with apt-get which
provides dependency resolution which dpkg does not.

Even if you install RVI with

   dpkg -i --ignore-depends=erlang rvi_0.5.0-1_armhf.deb

dpkg and apt-get will continue to complain that erlang is not installed
and refuse to install any package despite the fact that the dependencies
are fulfilled.

We can reinstantiate this in the future when Raspbian provides Erlang 18.2.

Signed-off-by: Rudolf J Streif <rudolf.streif@gmail.com>